+++ node/utils/readFile.ts
@@ -1,0 +1,7 @@
+import fs from 'fs'
+import path from 'path'
+
+const readFile = (filePath: string) =>
+  fs.readFileSync(path.resolve(__dirname, filePath), 'utf8')
+
+export default readFile